pyproject.toml
[project]
name = "agenticai-langgraph-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.13.3",
    "aiosignal==1.4.0",
    "aiosqlite>=0.22.1",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "asttokens==3.0.1",
    "attrs==25.4.0",
    "bs4>=0.0.2",
    "certifi==2026.1.4",
    "chardet==5.2.0",
    "charset-normalizer==3.4.4",
    "chromadb>=1.4.1",
    "colorama==0.4.6",
    "comm==0.2.3",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.19",
    "decorator==5.2.1",
    "executing==2.2.1",
    "faiss-cpu>=1.13.2",
    "frozenlist==1.8.0",
    "google-genai>=1.63.0",
    "greenlet==3.3.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.3",
    "idna==3.11",
    "ipykernel==7.1.0",
    "ipython==9.9.0",
    "ipython-pygments-lexers==1.1.1",
    "ipywidgets>=8.1.8",
    "jedi==0.19.2",
    "jiter>=0.13.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jupyter-client==8.8.0",
    "jupyter-core==5.9.1",
    "langchain==1.2.3",
    "langchain-chroma>=1.1.0",
    "langchain-classic==1.0.1",
    "langchain-community==0.4.1",
    "langchain-core==1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.7",
    "langchain-postgres>=0.0.17",
    "langchain-tavily>=0.2.17",
    "langchain-text-splitters==1.1.0",
    "langgraph==1.0.5",
    "langgraph-checkpoint==3.0.1",
    "langgraph-cli[inmem]>=0.4.14",
    "langgraph-prebuilt==1.0.5",
    "langgraph-sdk==0.3.2",
    "langsmith==0.6.2",
    "marshmallow==3.26.2",
    "matplotlib-inline==0.2.1",
    "multidict==6.7.0",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "numpy==2.4.1",
    "orjson==3.11.5",
    "ormsgpack==1.12.1",
    "packaging==25.0",
    "parso==0.8.5",
    "platformdirs==4.5.1",
    "prompt-toolkit==3.0.52",
    "propcache==0.4.1",
    "psutil==7.2.1",
    "psycopg[binary]>=3.3.3",
    "pure-eval==0.2.3",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pydantic-settings==2.12.0",
    "pygments==2.19.2",
    "pymupdf==1.26.7",
    "pypdf>=6.7.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "sentence-transformers>=5.2.0",
    "six==1.17.0",
    "sqlalchemy==2.0.45",
    "stack-data==0.6.3",
    "tenacity==9.1.2",
    "tornado==6.5.4",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typing-extensions==4.15.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uuid-utils==0.13.0",
    "wcwidth==0.2.14",
    "wikipedia>=1.4.0",
    "xxhash==3.6.0",
    "yarl==1.22.0",
    "zstandard==0.25.0",
]
